Closes the request stream message when the HTTP/2 stream is removed from the connection.
In the case of an abrupt HTTP/2 connection close, HTTP/2 streams will be removed from the connection before the client has sent an END_STREAM frame. Without handling this and closing the request stream message, there is no way for server logic to handle this error for e.g., cleanup actions, especially if they disable response timeout because the stream is intended to have infinite length.

Because clients are expected to send END_STREAM in normal circumstances, this case is handled as a ClosedSessionException.

Logic change originally proposed by @tobias-.

Fixes #971